iron teeth character trait (#6512)
# About the pull request
adds "iron teeth" trait for 1 point that allows marines to hold stuff in
their face slot with no fear of it falling out when they drop
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
i tried to larp as rambo and my combat knife fell out of my mouth and i
thought man you can only hold minor shit in your face slot would it be
cool if it was a little more useful so here we are
besides more character traits is good theres a shortage of them i think
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

got the trait... put a knife in my mouth... rested... hit myself with an
explosion to stun... knife was still in my mouth... no runtimes

</details>


# Changelog
:cl:
add: Adds "Iron teeth" trait.
/:cl:

---------

Co-authored-by: Contrabang <91113370+Contrabang@users.noreply.github.com>
Co-authored-by: Zonespace <41448081+Zonespace27@users.noreply.github.com>