Adds map name to game mode logging at round start. (#6551)
# About the pull request

Adds the name of the map to the round start log with the game mode.

# Explain why it's good for the game

This will make tracking map related data easier via logs and data
collection. Evidently we did not do this in a very obvious place before,
and the only way to know that a map was on while looking at logs
exclusively was through context clues.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
code: log the map name at round start
/:cl: