Adds chemical property encephalophrasive (#6567)
# About the pull request
Adds a chemical property which causes pain but allows the host to
communicate telepathically

Encephalophrasive is made by mixing Nervestimulating, and
Psychostimulating

OD causes severe brain damage
Critical OD causes paralysis

It's not a mystery that I'm a comics nerd, and it's touched upon
multiple times within the comics, and games that humans who have
exceptionally strong wills (or have enhanced willpower due to drugs) are
able to cause xenomorphs to 'falter' and pause for a moment before
acting. While telepathy doesn't necessarily do that, it has the
potential to do so.

Additionally, humans (in real life) already transmit brainwaves to
others and can percieve (at a very low level) other peoples brain waves,
so it's not entirely sci-fi nonsense.

# Explain why it's good for the game

Provides an extremely fluffy way for researchers to communicate with
xenomorphs and potentially eachother, as well as potentially making this
a "Wanted" chem in the same way as DNA Disintegrating is. (Though
probably to a significantly lesser degree)

# Testing Photographs and Procedure



<details>
<summary>Screenshots & Videos</summary>

https://youtu.be/qQ8JIavtHAA

</details>


# Changelog
:cl:
add: Adds the Encephalophrasive property for chemicals
del: Removed bit of description that falsely said that higher potency
affected clarity of messages
fix: Removes focusing from the combination to prevent conflict between
focusing and nerve-stimulating, thus preventing encephalophrasive from
forming
fix: Revises the second overdose to be for critical overdoses, as it was
intended to be
code: Swaps 'M' variables for 'chem_host'
code: rehauled how ability 'giving' is handled, and is now only
processed upon initial ingestion of the chemical instead of every tick
add: Adds a brief piece of flavor text upon initial ingestion.
/:cl:

---------

Co-authored-by: harryob <55142896+harryob@users.noreply.github.com>