USCM ERT Station | USS Kurtz (#6368)
# About the pull request

Adds a USCM ERT station that can be used for ERTs. The station consists
of the entire lower-deck and a tiny snap-shot of the upper deck based
off of the Sulaco.

The name of the ship is the 'USS Kurtz', this is based off of the
character from Joseph Conrad's novel 'Heart of Darkness', this is meant
to follow the criteria of USCM ships being named off of concepts, people
or places from his novels.

Currently only the CBRN ERT will use this.

# Explain why it's good for the game

Bit odd the CBRN were coming from a WY ship. More roleplay options as to
being rescued by a USCM warship, or for admin events

Nice Nostalgia to see the ole Sulaco back in action as well. 


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


</details>


# Changelog
:cl:
mapadd: The USS Kurtz has been added as a USCM affiliated ERT station.
The design off the USS Kurtz is based off of the USS Sulaco, from older
versions of CM-SS13. The USS Kurtz features the entire lower-deck and a
snapshot of the upper deck.
mapadd: The USCM CBRN ERT will use the USS Kurtz as their homebase. 
/:cl:

---------

Co-authored-by: Steelpoint <alexander.henley@hotmai.com>