Hivelord meson vision (#6570)
# About the pull request
Gives hivelords meson vision
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Makes it easier to see how what you are designing is coming together,
which should be a thing for a caste that is entirely focused on
building.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: Hivelords now have meson vision
add: Added a toggle for meson vision for hivelord and burrower
/:cl:

---------

Co-authored-by: harryob <55142896+harryob@users.noreply.github.com>