Adds a plantable UA flag in the CIC armoury (#6492)
# About the pull request

adds a plantable united americas flag in the CIC armoury.

![cic
armoury](https://github.com/cmss13-devs/cmss13/assets/17518895/8e821e3b-6b18-4fdf-a601-f0682b506611)

it takes 6 seconds to hoist up and to take down. raising the flag while
on harm intent will produce a unique warcry. if the flag is raised while
there are more than 14 nearby marines, a stronger rallying cry will
played with 40% of the marines automatically *warcry-ing. this has a 90
second cooldown.

the code should be relatively modular (i think?? i'm not very good at
coding and all the code is spaghetti-y), so adding in a UPP variant
should be a piece of cake. if some feels up to it, at least.

i am no spriter either, so they might not be super pretty but they're
!!animated!!

# Explain why it's good for the game

i don't know how to explain this besides it being pretty soulful and and
one of the few good ideas from the ideaguys forum. might be a pretty
decent way to coordinate marine pushes/holds? who knows


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>



https://github.com/cmss13-devs/cmss13/assets/17518895/5560bb9f-72a3-4805-8449-ac405e564092



https://github.com/cmss13-devs/cmss13/assets/17518895/b65817ab-69bf-4091-b6b7-6e955799e4b3



</details>


# Changelog
:cl:
add: The CIC armoury now contains a plantable flag of the United
Americas.
/:cl:

---------

Co-authored-by: Zonespace <41448081+Zonespace27@users.noreply.github.com>