Abomination Tutorial (#6442)
# About the pull request

"This'll be a nice and simple 2 hour PR" - Me, 6 hours ago


![image](https://github.com/cmss13-devs/cmss13/assets/41448081/9ee648df-4a48-4e94-a2ac-479e8a4a77a5)

Adds a tutorial for how to play abom. You cannot roll abom without
having completed its tutorial.

# Explain why it's good for the game

More tutorials are good, and encouraging players to know the basics of a
decently complicated and rare xeno before playing it is a good idea.

# Changelog
:cl:
add: Added a tutorial for xenomorph Abominations. You must complete the
tutorial before being able to roll for Abomination.
/:cl:

---------

Co-authored-by: John Doe <johndoe@gmail.com>
Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>