Splits the civ-only combat reporter gear into their own category. (#6676)
# About the pull request

Title

# Explain why it's good for the game

We get a lot of ahelps from first-time players of the combat
correspondant role, asking why they can't get these equipment pieces.
This should make it clearer.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: The Combat Correspondant's civilian equipment is now in it's own
category to make it easier to tell why you can't click it as the
military version.
/:cl: