Planted flag bugfix and warcry removal (#6737)
# About the pull request

- Fixes the flag being able to be duplicated
- The flag does a small amount of damage when hitting someone
- Item flag is indestructable
- Removes forced warcries

# Explain why it's good for the game

- Bugs are bad etc. etc.
- You are hitting someone with a big flagpole, it's fair that it does a
bit of damage.
- Stops the flag from getting annihilated by a stray nade or etc
- I talked with another maintainer about this, and I was given some fair
arguments about why forcing emotes is bad. In summary, on an RP server,
the player in charge of a character should ideally have as much control
over what the character does as possible. While this isn't the case for
things like brain damage causing forced movement, those are consequences
of negative things occurring to a character that serves as a hindrance
to being able to play. As such, you should try to minimize the amount of
times that forced emotes occur when not as a punishment, such as here.
- What hasn't changed:
- The flag planter still warcries
- Having >= 14 people still makes a custom warcry sound for the flag
planter.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

![image](https://github.com/user-attachments/assets/953b435f-05ac-438f-97f2-1fec511e1051)

</details>


# Changelog
:cl: Zonespace, Vile Beggar
add: Handheld planted flags are now indestructible and do a small amount
of damage on hit.
fix: Plantable flags can no longer be duplicated.
del: Planting a UA flag no longer forces anyone but the planter to
warcry.
/:cl:

---------

Co-authored-by: John Doe <johndoe@gmail.com>