Changes the tank's primary flamer to fire a 3-tile wide glob of high-combustion napalm (#6579)
# About the pull request

the tank's primary flamer weapon (the "DRG-N Offensive Flamer Unit") now
fires a glob of high-combustion napalm that spreads to a 3-tile wide
circle, somewhat similar to what it was before. it burns slightly hotter
than blue flame (45 vs 40 intensity). it also spawns a harmless smoke
cloud for cosmetic effect. my reasoning behind this is that the tank
does not have a way of extinguishing the flames on its own, and with how
quickly it can fire off flame globs it could potentially stall pushes
for a really, really long time. that kind of ruins the "offensive" part
of the weapon imo. high-combustion napalm burns out quickly which would
probably lessen this issue. the slightly higher damage output should be
offset by the fact that you can extinguish the flames in one roll.

~~i am also unsure of why this glob radius got removed. i couldn't find
anything relating to it's removal, nor if it was intended or not. it
might be worthwhile to lower the radius by one tile (imo this a bit too
much range) but for now i'm just reverting it to how it was before.~~

~~i at least think this is how it acted before (barring the short
duration, that's just something i changed). i'm just working off what i
have since i don't have any memories of the tank from back then.~~

<details>
<summary>old picture for reference</summary>


![image](https://github.com/cmss13-devs/cmss13/assets/17518895/8b21eb74-d7fd-4ef1-a701-0580786e51bc)

</details>

# Explain why it's good for the game
currently the primary flamer just fires 1 tile of blue flame which is
very underwhelming for what is supposed to be a primary weapon for the
tank. it might need a few tweaks but with these changes it will be a
very potent weapon, fit for being a proper primary.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>



https://github.com/cmss13-devs/cmss13/assets/17518895/a04aaf55-7ab8-4446-82c3-cbe76b6c696f



Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: The DRG-N Offensive Flamer Unit now fires a 3-tile wide glob of
high-combustion napalm.
/:cl: