Fixes utensils attacking when trying to feed while full (#6679)
# About the pull request

Using a utensil with food on a target that's not hungry no longer stabs
them.

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game

You have a fork with some food, you're on help intent, you use it on
yourself... and you stab yourself to the chorus of "You don't feel like
eating more right now."
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: no longer stab with utensil when trying to feed while target full
/:cl: