Removes facial disfigurement (#6651)
# About the pull request

Removes facial disfigurement.

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game

Facial disfigurement is a roleplay tool, one which doesn't really have a
place in CM due to its rules around messing with other people's IDs. I
have never seen any sort of impersonation plot or deeper engagement from
facial disfigurement. The most I've seen is "OH GOD MY FACE!" and then
the marine gets it fixed, or ignores it.

So why not leave it as-is? It causes no harm. While leaving it is
harmless it is just one more aspect of "standard" SS13 that doesn't fit
with CM. One more piece of code that has to be maintained through
changes and refactors. One more thing to consider when trying to
understand what the code does. In the same vein as the removal of
walk/run mode and aim mode.
# Testing Photographs and Procedure
Boots without issue. Facial damage, yautja skinning, and var editing
still work minus the facial disfigurement.


# Changelog
:cl:
del: Removed facial disfigurement
/:cl: