Fixes item actions not working properly (#6756)
# About the pull request
Calling the parent of all actions
(https://github.com/cmss13-devs/cmss13/pull/6442) caused a lot of bugs,
this fixes that.

Closes https://github.com/cmss13-devs/cmss13/issues/6755
Closes https://github.com/cmss13-devs/cmss13/issues/6753
Closes #6732 
Closes https://github.com/cmss13-devs/cmss13/pull/6742

# Explain why it's good for the game

Bugs are bad

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Tested on practically every object that has a use-in-hand functionality
with an action button.

</details>


# Changelog
:cl:
fix: Fixed a large amount of objects activating erroneous functionality
(like an armor light) when using action buttons tied to that object.
/:cl:

Co-authored-by: John Doe <johndoe@gmail.com>