HPR Starts With Bipod Attached (#6715)
# About the pull request

See title. The HPR will now spawn in with a Bipod attached to it. 

# Explain why it's good for the game

The concept of the HPR having access to a unique auto-fire function,
with a very high ROF, is not at all advertised to the average player,
this is a concept wherein many people are left unaware of this function
until it is revealed to them by a third party. This is common testimony
from those who discuss the HPR and its bipod function.

Forcing the HPR to spawn with a bipod attached to it, will resolve the
manner as it will expose far more people to the ability for the weapon
to fire on auto-fire and thus reveal an entirely new and unique way to
make usage of the firearm to people who may never have discovered it.

Considering the bipod is a fairly common attachment that is not often
selected, and the Almayer has a very limited supply of HPRs, I do not
believe this will have any real impact on attachment availability.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: The Heavy Pulse Rifle (M41AE2) will now start with a pre-attached
bipod. To better inform players of its unique auto-fire function.
/:cl:

---------

Co-authored-by: Steelpoint <alexander.henley@hotmai.com>