Fix quadtree AABB issue (#6782)
# About the pull request

Quadtree AABB testing now correctly uses the Y bounds to test if the Y
coordinate is in range.

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game

Bug bad. Shouldn't use the X coordinate to determine if you're in bounds
of the Y coordinate. Fortunate that so many things are squares that it
didn't cause an issue.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: sounds & motion detectors should be more reliable
/:cl: