Adds Urban Sprite For Ghillie Suit (#6563)
# About the pull request

Replaces the Green ghillie sprites for the Sniper/AMR into a more Urban
sprite, using the Elite Mercenary Sprites akin to Forecon, for urban
maps like Chances Claim or Fiorina.

# Explain why it's good for the game

Very odd the Sniper/AMR are using Green on a grey map for camo. This
makes it look appropriate, and precedent the USCM uses this scheme in
forecon.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

![Screenshot 2024-06-26 04 48
39](https://github.com/cmss13-devs/cmss13/assets/6595389/3fb6b07c-c304-47e4-9a4b-56b8b7d08b6a)


</details>


# Changelog
:cl:
imageadd: The Sniper and Anti-Material Rifle Specialists will now use an
urban camouflage sprite for their ghillie suit instead of a green
camouflage.
/:cl:

Co-authored-by: Steelpoint <alexander.henley@hotmai.com>