Adds same hats available to Staff Officers to the ASO vendor (#6716)
# About the pull request

ASO vendor now has a few more options for hats in their loadout. The
same options that are already provided to SOs. Specifically it gives
them access to the following hats that they previously didn't have as an
option:

- Beret, green
- Beret, tan
- Officer cap
- Service Peaked Cap

The only hat they had access to previously was the "patrol cap" and this
remains as an option for them to select.

# Explain why it's good for the game

More options for customizing appearance are good when it makes sense and
fits within uniform SOP. Given that ASO is a higher rank than SO and is
a shipside role with only a very niche case in which they would ever
deploy groundside, it makes sense to let them have a bit more
flexibility on their uniform appearance similar to other officers on the
ship.

# Testing Photographs and Procedure
<details>

Tested by building the code after this change and observing the ASO
vendor to see that new options appear. Confirmed that it only lets you
select one option and doesn't conflict with also letting you select a
helmet (same as how XO is handled).

<summary>Screenshots & Videos</summary>


![image](https://github.com/user-attachments/assets/bd425a60-8acc-4c12-afb3-b15aa3c97c84)

Screenshot of vendor after selecting a single option for a hat.

</details>


# Changelog
:cl:
add: more options for hats to ASO vendor to bring it in line with what
SO has available
/:cl: