capitalize W in workshop for ordnance workshop area (#6813)
# About the pull request

capitalize W in workshop for ordnance workshop area

# Explain why it's good for the game

mostly so the display on the vitals monitor computer is congruent with
the titles for other locations but this will also change the name as it
appears when viewing the actual area itself


# Testing Photographs and Procedure
tested.

# Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
label your changes in the changelog. Please note that maintainers freely
reserve the right to remove and add tags should they deem it
appropriate. You can attempt to finagle the system all you want, but
it's best to shoot for clear communication right off the bat. -->
<!-- If you add a name after the ':cl', that name will be used in the
changelog. You must add your CKEY after the CL if your GitHub name
doesn't match. Maintainers freely reserve the right to remove and add
tags should they deem it appropriate. -->

:cl:
spellcheck: Fixed capitalization of ordnance workshop area
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->

Co-authored-by: nefarious <72267018+nefarious6th@users.noreply.github.com>