Another NERF to heavy revolver ammo (#6775)
# About the pull request
https://github.com/cmss13-devs/cmss13/pull/4706
So this PR replaced knockdown with slow+knockback, but unlike the
knockdown proc the slow proc didn't check for the mob size, so you
became able to slow down T3s and even Queen with heavy revolver ammo,
which clearly wasn't intended. This wasn't noticed until recently.

# Explain why it's good for the game
Free slows with no telegraph on T3s is not something we need in the
game.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: ihatethisengine2
balance: heavy revolver ammo cannot slowdown t3s anymore.
/:cl: