Stops the queen from resting while on an ovipositor (#6835)
# About the pull request

stops the queen from resting while on ovi.
Fixes #6685

# Explain why it's good for the game

it's a bug. you should not be able to rest on ovi as it doesn't make
sense and only ends up making half of your abilities unusable.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

kinda hard to show this but here's me mashing the rest verb with my
mouse


https://github.com/user-attachments/assets/103f3f0f-966b-4084-8888-b948fbc2bee0



</details>


# Changelog
:cl:
fix: You can no longer rest while on ovi as a Queen, which made half of
your abilities unusable if you did so.
/:cl: