Stop phones from going in lockers/crates (#6856)
# About the pull request

stops phones from going in lockers/crates, which prevents players from
exploiting the phone drag feature.
fixes #6201

i tried making it so the phones would bust out of the crate, but it
would look really odd as reset_tether() would get called and the phone
would magically teleported and become attached to the backpack. i
couldn't find a way to make it smoothly pop out of the crate, and i
think this is a decent enough alternative, as i don't know if a single
situation where a player would need to store a phone.

this also stops the phone tether from becoming destroyed if you put it
in an anchored closet.

# Explain why it's good for the game

fixes a (funny) exploit and another niche issue


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.


https://github.com/user-attachments/assets/addca004-27b1-4a8d-887e-c7600d7851a2



</details>


# Changelog
:cl:
fix: Phones can no longer be stored in closets/crates, which prevents
players from dragging crates with their phones and other weird issues
that can occur.
/:cl: