Fixes and Readjusts Larva Surgery  (#6864)
# About the pull request

Currently, the way the larva surgery was written, using the hemostat for
the final larva removal step was **SLOWER** than using ghetto tools such
as the wirecutters and fork. This PR aims to fix the issue by making the
hemostat the best tool, whilst keeping the original time it took to do
so, and readjusts the niche "using one's hands" in an emergency to be
the fastest way at the cost of both the doctor's and patient's health.

People used to ghetto surgery on larva will notice it being much slower.

# Explain why it's good for the game

It makes no sense for wirecutters and the fork to be faster than using
the intended tool. Having the option to make an emergency procedure with
one's hands gives doctors a satisfying way of saving their patients from
impending chestburst death.

# Testing Photographs and Procedure

N/A

# Changelog

:cl: MarpleJones
fix: Using the hemostat on the final larva surgery step is now better
than using wirecutters or the fork.
add: The final larva surgery step will now apply organ damage when done
barehanded. Barehanded is faster than using tools by a couple of
seconds, at the cost of the doctor's and the patient's health.
add: Adds an acidsplash sound effect to doing the barehanded step.
/:cl:

---------

Co-authored-by: MarplePhonks <marplejones@gmail.com>
Co-authored-by: Contrabang <91113370+Contrabang@users.noreply.github.com>
Co-authored-by: harryob <55142896+harryob@users.noreply.github.com>