Prevents tunnels from being dug under the LZ Sentry holders. (#6871)
# About the pull request
Addresses #6773 by making the dig tunnel ability check for the LZ|
sentry holders.
Wording could probably be better.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Presumably it isn't intended to be able to hide FOB tunnels completely
under the LZ sentries.
So this rectifies this being able to happen.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: MistChristmas
fix: Prevents tunnels under LZ Sentries
/:cl: