Stethoscopes and penlights - An attempt on usefulness (Also, flashlight flashing fixed) (#6691)
# About the pull request

The main purpose of this PR is to add functionality to both the
stethoscope and the penlight, with the fixing of the "flash eyes with
flashlight" mechanism being only a conterminous addon.

Special thanks to Drathek for assisting with coding this, without his
help I wouldn't go anywhere.

# Explain why it's good for the game

Adds more detail to the game and allows the player to have more options
regarding the diagnosis of their patients. Even though this PR is not
meant to replace any type of current diagnosis methods, it is thought to
allow the possibility of these items being used if the player wish, or
if the situation they are in requires. It is also thought to actually
give some purpose to both items, instead of only being decorative
(basically).

About the "flash eyes with flashlight" mechanism, it just have been
repaired because of its relation with the functionality of the penlight
itself. (Remember, only the penlight can be used to do diagnosis,
generic flashlights will only allow to temporarily blind the person you
are interacting with, and will not give you any medical information).

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/user-attachments/assets/7f303857-7746-4258-9740-f37b01b3033e)


![image](https://github.com/user-attachments/assets/1f3ec276-b44d-4e4d-a708-3361eb008c62)


![image](https://github.com/user-attachments/assets/ae172eaf-52bf-4208-b548-9fbdd3b86d21)


![image](https://github.com/user-attachments/assets/57017a46-3eea-492c-80d8-0c9a18dfb36b)


![image](https://github.com/user-attachments/assets/93508086-ae32-449e-879a-8acc816bd354)

</details>


# Changelog
:cl:
add: Adds penlights and stethoscopes to nurses and corpsmen, they can
use them to reliably check if someone have specific types of organ
damage and how much, without the need of other apparatus. Adds a new
organ.status called "LITTLE_BRUISED", used by the stethoscope and
penlight to not give false positives when diagnosticating someone
(saying that they are healthy, when they actually have 9 heart damage).
fix: Fixed an outdated check that made the entire "flash eyes with
flashlight" mechanic don't work at all. Also updated it to check some
other things for ease of logic.
spellcheck: Fixed typos on the abandoned "flash eyes with flashlight"
code.
code: Added new functionalities for both the stethoscope and penlight:
The stethoscopes can now be used to check the condition of both heart
and lungs individually by aiming the chest, letting you be able to check
if the organs are LITTLE_BRUISED (Have a damage equal or above 1),
BRUISED (Have a damage equal or above 10), BROKEN (have a damage equal
or above 30) or "HEALTHY" (any damage below 1). And the penlight, that
can be used the same way to check the condition of brain and eyes by
aiming the eyes of the person you are interacting with. Also with the
fixing mentioned above, you can now flash people's eyes again! Everytime
you check the condition of someone's eyes, you will also flash them.
Replaced one letter vars in the flashlight code and on the stethoscope
section of the ties code.
imageadd: Adds new "in_hand" icons for penlights and stethoscopes, both
will now appear on your hands when used. Penlights have both "off" and
"on" icons, fancy stuff!
/:cl:

---------

Co-authored-by: Zonespace <41448081+Zonespace27@users.noreply.github.com>
Co-authored-by: kiVts <48099872+kiVts@users.noreply.github.com>